+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 All notable changes to this project will be documented in this file.  
 
 - 2021-11-01: formal software release  
+- 2021-12-08: add a commandline tool convert_mtx.py to convert cellranger output files to Tres input format.